move content of assets.ts to experimental-assets.ts
